+++ src/state_feed_backward_cpu.cpp
@@ -3,7 +3,7 @@
 // Description:  CPU version for backward pass for hidden state
 // Authors:      Luong-Ha Nguyen & James-A. Goulet
 // Created:      May 18, 2022
-// Updated:      August 31, 2022
+// Updated:      March 01, 2023
 // Contact:      luongha.nguyen@gmail.com & james.goulet@polymtl.ca
 // Copyright (c) 2022 Luong-Ha Nguyen & James-A. Goulet. Some rights reserved.
 ///////////////////////////////////////////////////////////////////////////
@@ -120,6 +120,34 @@ Args:
             delta_S[col + z_delta_pos] = zeroPad;
         } else {
             delta_S[col + z_delta_pos] = tmp / Sz[col + z_pos];
+        }
+    }
+}
+
+////////////////////////////////////////////////////////////////////////////////
+/// REMAX
+////////////////////////////////////////////////////////////////////////////////
+void delta_z_y_check_cpu(std::vector<float> &mu_a, std::vector<float> &var_a,
+                         std::vector<float> &cov_y_y_check,
+                         std::vector<float> &y, std::vector<float> &var_noise,
+                         int no, int B, int z_pos,
+                         std::vector<float> &delta_mu_zy_check,
+                         std::vector<float> &delta_var_zy_check)
+/*Compute updating quantities for \check{y}*/
+{
+    float tmp = 0, zero_pad = 0;
+    int col;
+    for (int i = 0; i < B; i++) {
+        for (int j = 0; j < no; j++) {
+            col = i * no + j;
+            tmp = cov_y_y_check[col] / (var_a[col + z_pos] + var_noise[col]);
+            if (isinf(tmp) || isnan(tmp)) {
+                delta_mu_zy_check[col] = zero_pad;
+                delta_var_zy_check[col] = zero_pad;
+            } else {
+                delta_mu_zy_check[col] = tmp * (y[col] - mu_a[col + z_pos]);
+                delta_var_zy_check[col] = -tmp * cov_y_y_check[col];
+            }
         }
     }
 }
@@ -888,6 +916,30 @@ void output_delta_mz_Sz_cpu(Network &net, NetState &state, Obs &obs,
     }
 }
 
+void remax_output_delta_z_cpu(Network &net, NetState &state, Obs &obs,
+                              DeltaState &d_state)
+/*Compute updating quantities of hidden states for softmax layer*/
+{
+    int no = net.nodes.back();
+    int B = net.batch_size;
+    int z_pos = net.z_pos.back();
+    // Covariance between m and \check{a}
+    compute_cov_m_a_check_cpu(state.remax.var_log, state.remax.cov_log_logsum,
+                              state.remax.mu_m, no, B,
+                              state.remax.cov_m_a_check);
+
+    // Covariance between m and a
+    compute_cov_m_a_cpu(state.remax.cov_m_a_check, state.ma, state.remax.var_m,
+                        state.Sz, state.remax.J_m, z_pos, no, B,
+                        state.remax.cov_m_a);
+
+    // Updating quantities for hidden states
+    delta_z_y_check_cpu(state.ma, state.Sa, state.remax.cov_m_a, obs.y_batch,
+                        obs.V_batch, no, B, z_pos, d_state.delta_mz,
+                        d_state.delta_Sz);
+    int check = 1;
+}
+
 void update_output_hidden_states_cpu(Network &net, NetState &state, Obs &obs,
                                      DeltaState &d_state)
 /*Compute updated quantities for the output layer's hidden state
@@ -901,8 +953,11 @@ void update_output_hidden_states_cpu(Network &net, NetState &state, Obs &obs,
 {
     if (net.is_output_ud) {
         if (net.noise_type.compare("homosce") != 0 &&
-            net.noise_type.compare("heteros") != 0) {
+            net.noise_type.compare("heteros") != 0 &&
+            (net.activations.back() != net.act_names.remax)) {
             output_delta_mz_Sz_cpu(net, state, obs, d_state);
+        } else if (net.activations.back() == net.act_names.remax) {
+            remax_output_delta_z_cpu(net, state, obs, d_state);
         } else {
             output_delta_mz_Sz_with_noise_inferenece_cpu(state, net, obs,
                                                          d_state);